app/requirements.txt
Flask==2.0.3
google-cloud-datastore==2.4.0
google-auth==2.5.0
requests==2.27.1
gunicorn==20.1.0
google-cloud-texttospeech==2.11.0